# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SLM Fine-Tuning Platform",
    "description": "Backend-only PoC for an automated Small Language Model fine-tuning platform. Generates synthetic data via OpenRouter, fine-tunes ≤3B models with Unsloth + QLoRA, tracks every run in MLflow, and serves results via Ollama.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/datasets/upload": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Upload a ready-to-train dataset (JSONL or JSON)",
        "operationId": "upload_dataset_api_v1_datasets_upload_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_api_v1_datasets_upload_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeedUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/upload-seed": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Upload seed examples (JSONL or JSON)",
        "operationId": "upload_seed_dataset_api_v1_datasets_upload_seed_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_seed_dataset_api_v1_datasets_upload_seed_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeedUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/generate": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Generate a synthetic dataset via OpenRouter",
        "operationId": "generate_dataset_api_v1_datasets_generate_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/SDGRequestWithSeed"
                  },
                  {
                    "$ref": "#/components/schemas/SDGRequestDescriptionOnly"
                  }
                ],
                "discriminator": {
                  "propertyName": "sdg_mode",
                  "mapping": {
                    "with_seed": "#/components/schemas/SDGRequestWithSeed",
                    "description_only": "#/components/schemas/SDGRequestDescriptionOnly"
                  }
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SDGJobAcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "List datasets (optionally filter by project)",
        "operationId": "list_datasets_api_v1_datasets_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DatasetResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get a dataset by id",
        "operationId": "get_dataset_api_v1_datasets__dataset_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "datasets"
        ],
        "summary": "Rename a dataset",
        "operationId": "update_dataset_api_v1_datasets__dataset_id__patch",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "datasets"
        ],
        "summary": "Delete a dataset",
        "operationId": "delete_dataset_api_v1_datasets__dataset_id__delete",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/preview": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Preview the first N rows of a dataset",
        "operationId": "preview_dataset_api_v1_datasets__dataset_id__preview_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/download": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Download the raw dataset file (JSONL)",
        "operationId": "download_dataset_api_v1_datasets__dataset_id__download_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/download-url": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Mint a presigned MinIO URL for the dataset's stored object",
        "operationId": "get_dataset_download_url_api_v1_datasets__dataset_id__download_url_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          },
          {
            "name": "disposition",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "attachment",
                "inline"
              ],
              "type": "string",
              "default": "attachment",
              "title": "Disposition"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDownloadUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/cancel": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Cancel a running SDG generation job (idempotent)",
        "operationId": "cancel_dataset_api_v1_datasets__dataset_id__cancel_post",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Cancel Dataset Api V1 Datasets  Dataset Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_dataset_api_v1_datasets_upload_post": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "task_type",
          "file"
        ],
        "title": "Body_upload_dataset_api_v1_datasets_upload_post"
      },
      "Body_upload_seed_dataset_api_v1_datasets_upload_seed_post": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "task_type",
          "file"
        ],
        "title": "Body_upload_seed_dataset_api_v1_datasets_upload_seed_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SDGJobAcceptedResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "description": "Celery task id; same as the WebSocket channel suffix."
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id",
            "description": "Placeholder dataset row created at submission."
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "pending"
          },
          "websocket_url": {
            "type": "string",
            "title": "Websocket Url",
            "description": "Convenience URL for the per-job progress stream.",
            "examples": [
              "ws://localhost:8000/ws/jobs/celery-task-id-here"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "dataset_id",
          "websocket_url"
        ],
        "title": "SDGJobAcceptedResponse",
        "description": "202 Accepted body — the dataset row exists but has 0 rows until completion."
      },
...
      "SDGRequestWithSeed": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id",
            "description": "Owning project."
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "task_description": {
            "type": "string",
            "minLength": 10,
            "title": "Task Description",
            "description": "Natural-language description of what the model should learn."
          },
          "num_samples": {
            "type": "integer",
            "maximum": 10000.0,
            "exclusiveMinimum": 0.0,
            "title": "Num Samples",
            "description": "How many synthetic rows to generate."
          },
          "holdout_size": {
            "type": "integer",
            "maximum": 2000.0,
            "minimum": 0.0,
            "title": "Holdout Size",
            "description": "Extra rows generated beyond `num_samples`, persisted as a separate child Dataset (linked via parent_dataset_id) for hold-out evaluation. Set to 0 to disable. Stratified by label (classification) / tool name (tool_calling); random for QA.",
            "default": 100
          },
          "temperature": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "default": 0.9
          },
          "dataset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Name",
            "description": "Display name for the resulting dataset; defaults to project + timestamp."
          },
          "holdout_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holdout Name",
            "description": "Display name for the holdout dataset; defaults to '<train dataset name>-hold-out'."
          },
          "sdg_mode": {
            "type": "string",
            "const": "with_seed",
            "title": "Sdg Mode",
            "default": "with_seed"
          },
          "seed_dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Seed Dataset Id",
            "description": "ID of a previously-uploaded seed dataset (POST /api/v1/datasets/upload-seed). The dataset's source must be DatasetSource.SEED and task_type must match this request's task_type. For QA + PDF, the dataset's metadata must contain a `pdf_uri`."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "task_type",
          "task_description",
          "num_samples",
          "seed_dataset_id"
        ],
        "title": "SDGRequestWithSeed",
        "description": "Generate by extrapolating from a previously-uploaded seed dataset.\n\nUpload the seed first via `POST /api/v1/datasets/upload-seed`, then\npass the returned `dataset_id` here as `seed_dataset_id`.",
        "examples": [
          {
            "holdout_size": 50,
            "num_samples": 200,
            "project_id": "00000000-0000-0000-0000-000000000001",
            "sdg_mode": "with_seed",
            "seed_dataset_id": "00000000-0000-0000-0000-000000000099",
            "task_description": "Answer questions about our 30-day return policy",
            "task_type": "qa",
            "temperature": 0.9
          }
        ]
      },
...
      "SeedUploadResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "num_samples": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Num Samples",
            "description": "Number of canonicalised rows persisted to MinIO. 0 for PDF uploads — Q&A pairs come from the SDG generator later."
          },
          "invalid_rows": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Invalid Rows",
            "description": "Indexes of rows that failed validation (if any were tolerated)."
          },
          "format_detection": {
            "$ref": "#/components/schemas/FormatDetectionReport",
            "description": "Audit trail of the Format Detection pass. `ran=False` when the seed was already canonical or for PDF uploads."
          },
          "pdf_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Uri",
            "description": "S3 URI of the persisted PDF (only set for QA + PDF uploads)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset_id",
          "task_type",
          "num_samples",
          "format_detection"
        ],
        "title": "SeedUploadResponse",
        "description": "Response for `POST /api/v1/datasets/upload-seed` (multipart upload).\n\nPhase 9: extended with `format_detection` (the audit trail of the\nschema-mapping pass) and `pdf_uri` (set only for QA + PDF uploads)."
      },
...
    }
  },
...
}